Implemented `KipperSemanticChecker` and `KipperTypeChecker` as explained in #124
